resources/test/tests/functional/precondition-failed.html

<!DOCTYPE HTML>
<html>
<head>
<meta name="variant" content="?keep-promise">
<title>assert_precondition</title>
<script src="../../variants.js"></script>
</head>
<body>
<h1>assert_precondition</h1>
<div id="log"></div>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
  test(function(t) {
    assert_precondition(false, "custom message");
  }, 'test (synchronous)');

  async_test(function(t) {
    t.step_timeout(function() {
      assert_precondition(false);
    }, 0);
  }, 'async_test');

  promise_test(function(t) {
    assert_precondition(false);

    return new Promise(function() {});
  }, 'promise_test');
</script>
<script type="text/json" id="expected">
{
  "summarized_status": {
    "status_string": "OK",
    "message": null
  },
  "summarized_tests": [
    {
      "status_string": "PRECONDITION_FAILED",
      "name": "async_test",
      "message": "Test skipped",
      "properties": {}
    },
    {
      "status_string": "PRECONDITION_FAILED",
      "name": "promise_test",
      "message": "Test skipped",
      "properties": {}
    },
    {
      "status_string": "PRECONDITION_FAILED",
      "name": "test (synchronous)",
      "message": "custom message",
      "properties": {}
    }
  ],
  "type": "complete"
}
</script>
</body>
</html>
